src/database/requirements.txt
django>=1.11.10
djangorestframework>=3.7.7
markdown>=2.6.11
django-filter>=1.1.0
djangorestframework-jwt>=1.11.0
django-cors-headers>=2.1.0
